+++ pallets/roles/src/impls.rs
@@ -17,7 +17,8 @@
 use super::*;
 use crate::offences::ValidatorOffence;
 use frame_support::{pallet_prelude::DispatchResult, traits::OneSessionHandler};
-use pallet_staking::{CurrentEra, ErasRewardPoints};
+use log::log;
+use pallet_staking::{CurrentEra, ErasRewardPoints, ForceEra, Forcing};
 use sp_runtime::{
 	traits::{CheckedDiv, Convert},
 	Perbill,
@@ -106,434 +107,6 @@ impl<T: Config> RolesHandler<T::AccountId> for Pallet<T> {
 	}
 }
 
-/// Functions for the pallet.
-impl<T: Config> Pallet<T> {
-	/// Validate updated profile for the given account.
-	/// This function will validate the updated profile for the given account by
-	/// checking if the account has any active jobs for the removed roles. If the
-	/// account has any active jobs for the removed roles, then it will return
-	/// the error `RoleCannotBeRemoved`.
-	///
-	/// # Parameters
-	/// - `account`: The account ID of the validator.
-	/// - `updated_profile`: The updated profile.
-	pub(crate) fn validate_updated_profile(
-		account: T::AccountId,
-		updated_profile: Profile<T>,
-	) -> DispatchResult {
-		let current_ledger = Self::ledger(&account).ok_or(Error::<T>::NoProfileFound)?;
-		let active_jobs: Vec<(RoleType, JobId)> = T::JobsHandler::get_active_jobs(account.clone());
-		// Check if the account has any active jobs for the removed roles.
-		let removed_roles = current_ledger.profile.get_removed_roles(&updated_profile);
-		if !removed_roles.is_empty() {
-			for role in removed_roles {
-				for job in active_jobs.clone() {
-					let role_type = job.0;
-					if role_type == role {
-						return Err(Error::<T>::RoleCannotBeRemoved.into())
-					}
-				}
-			}
-		};
-
-		// Get all roles for which there are active jobs
-		let roles_with_active_jobs: Vec<RoleType> =
-			active_jobs.iter().map(|job| job.0).fold(Vec::new(), |mut acc, role| {
-				if !acc.contains(&role) {
-					acc.push(role);
-				}
-				acc
-			});
-
-		// If there are active jobs, changing a current independent profile to shared profile
-		// is allowed if and only if the shared restaking amount is at least as much as the
-		// sum of the restaking amounts of the current profile. This is because we require
-		// the total amount staked to only increase or remain the same across active roles.
-		if updated_profile.is_shared() && current_ledger.profile.is_independent() {
-			if active_jobs.len() > 0 {
-				let mut active_role_restaking_sum = Zero::zero();
-				for role in roles_with_active_jobs.iter() {
-					let current_role_restaking_amount = current_ledger
-						.profile
-						.get_records()
-						.iter()
-						.find_map(|record| if record.role == *role { record.amount } else { None })
-						.unwrap_or_else(|| Zero::zero());
-					active_role_restaking_sum += current_role_restaking_amount;
-				}
-
-				ensure!(
-					updated_profile.get_total_profile_restake() >= active_role_restaking_sum,
-					Error::<T>::InsufficientRestakingBond
-				);
-			}
-		}
-
-		// Changing a current shared profile to an independent profile is allowed if there are
-		// active jobs as long as the stake allocated to the active roles is at least as much as
-		// the shared profile restaking amount. This is because the shared restaking profile for an
-		// active role is entirely allocated to that role (as it is shared between all selected
-		// roles). Thus, we allow the user to change to an independent profile as long as the
-		// restaking amount for the active roles is at least as much as the shared restaking amount.
-		if updated_profile.is_independent() && current_ledger.profile.is_shared() {
-			// For each role with an active job, ensure its stake is greater than or equal to the
-			// existing ledger's shared restaking amount.
-			for role in roles_with_active_jobs.iter() {
-				let updated_role_restaking_amount = updated_profile
-					.get_records()
-					.iter()
-					.find_map(|record| if record.role == *role { record.amount } else { None })
-					.unwrap_or_else(|| Zero::zero());
-				ensure!(
-					updated_role_restaking_amount >=
-						current_ledger.profile.get_total_profile_restake(),
-					Error::<T>::InsufficientRestakingBond
-				);
-			}
-
-			return Ok(())
-		}
-		// For each role with an active job, ensure its stake is greater than or equal to the
-		// existing ledger's restaking amount for that role. If it's a shared profile, then the
-		// restaking amount for that role is the entire shared restaking amount.
-		let min_restaking_bond = MinRestakingBond::<T>::get();
-		for record in updated_profile.clone().get_records() {
-			match updated_profile.clone() {
-				Profile::Independent(_) =>
-					if roles_with_active_jobs.contains(&record.role) {
-						ensure!(
-							record.amount.unwrap_or_default() >= min_restaking_bond,
-							Error::<T>::InsufficientRestakingBond
-						);
-						ensure!(
-							record.amount.unwrap_or_default() >=
-								current_ledger.restake_for(&record.role),
-							Error::<T>::InsufficientRestakingBond
-						);
-					},
-				Profile::Shared(profile) =>
-					if roles_with_active_jobs.contains(&record.role) {
-						ensure!(
-							profile.amount >= current_ledger.profile.get_total_profile_restake(),
-							Error::<T>::InsufficientRestakingBond
-						);
-					},
-			}
-		}
-		Ok(())
-	}
-
-	/// Check if account can chill, unbond and withdraw funds.
-	///
-	/// # Parameters
-	/// - `account`: The account ID of the restaker.
-	///
-	/// # Returns
-	/// Returns boolean value.
-	pub(crate) fn can_exit(account: T::AccountId) -> bool {
-		let assigned_roles = AccountRolesMapping::<T>::get(account);
-		if assigned_roles.is_empty() {
-			// Role is cleared, account can chill, unbond and withdraw funds.
-			return true
-		}
-		false
-	}
-
-	/// Calculate max restake amount for the given account.
-	///
-	/// # Parameters
-	/// - `total_stake`: Total stake of the restaker
-	///
-	/// # Returns
-	/// Returns the max restake amount.
-	pub(crate) fn calculate_max_restake_amount(total_stake: BalanceOf<T>) -> BalanceOf<T> {
-		// User can restake max 50% of the total stake
-		T::MaxRestake::get() * total_stake
-	}
-
-	/// Calculate slash value for restaked amount
-	///
-	/// # Parameters
-	/// - `slash_fraction`: Slash fraction of total-stake
-	/// - `total_stake`: Total stake of the validator
-	///
-	/// # Returns
-	/// Returns the slash value
-	pub(crate) fn calculate_restake_slash_value(
-		slash_fraction: Perbill,
-		total_stake: BalanceOf<T>,
-	) -> BalanceOf<T> {
-		// Slash value for given slash fraction
-		slash_fraction * total_stake
-	}
-
-	/// Report offence for the given restaker.
-	/// This function will report restakers for committing offence.
-	///
-	/// # Parameters
-	/// - `offence_report`: The offence report.
-	///
-	/// # Returns
-	///
-	/// Returns Ok() if restaker offence report is submitted successfully.
-	pub(crate) fn report_offence(
-		offence_report: ReportRestakerOffence<T::AccountId>,
-	) -> sp_runtime::DispatchResult {
-		let offenders = offence_report
-			.clone()
-			.offenders
-			.into_iter()
-			.enumerate()
-			.filter_map(|(_, id)| {
-				// Get validator id from account id.
-				let id = <<T as Config>::ValidatorSet as ValidatorSet<
-					<T as frame_system::Config>::AccountId,
-				>>::ValidatorIdOf::convert(id)?;
-				<T::ValidatorSet as ValidatorSetWithIdentification<T::AccountId>>::IdentificationOf::convert(
-					id.clone(),
-				).map(|full_id| (id, full_id))
-			})
-			.collect::<Vec<IdentificationTuple<T>>>();
-
-		let offence = ValidatorOffence {
-			session_index: offence_report.session_index,
-			validator_set_count: offence_report.validator_set_count,
-			offenders,
-			role_type: offence_report.role_type,
-			offence_type: offence_report.offence_type,
-		};
-		let _ = T::ReportOffences::report_offence(sp_std::vec![], offence.clone());
-		// Update and apply slash on ledger for all offenders
-		let slash_fraction = offence.slash_fraction(offence.validator_set_count);
-		for offender in offence_report.offenders {
-			let mut profile_ledger =
-				<Ledger<T>>::get(&offender).ok_or(Error::<T>::NoProfileFound)?;
-			let staking_ledger =
-				pallet_staking::Ledger::<T>::get(&offender).ok_or(Error::<T>::NotValidator)?;
-			let slash_value =
-				Self::calculate_restake_slash_value(slash_fraction, staking_ledger.total);
-			// apply slash
-			profile_ledger.total = profile_ledger.total.saturating_sub(slash_value);
-			Self::update_ledger(&offender, &profile_ledger);
-		}
-
-		Ok(())
-	}
-
-	/// Update the ledger for the given stash account.
-	///
-	/// # Parameters
-	/// - `restaker`: The stash account ID.
-	/// - `ledger`: The new ledger.
-	pub(crate) fn update_ledger(restaker: &T::AccountId, ledger: &RestakingLedger<T>) {
-		<Ledger<T>>::insert(restaker, ledger);
-	}
-
-	/// This function retrieves the current era and the total rewards available for distribution
-	/// from the runtime configuration. It then computes rewards for active validators and
-	/// validators based on their role types. After computing rewards, it adds individual reward
-	/// points to each eligible validator and records them in the staking pallet ErasRewardPoints
-	/// storage for the current era.
-	///
-	/// The total reward is calculated as follows:
-	/// 1. Fifty percent of the total rewards is divided among all "active" validators. Active
-	///    validators
-	/// are those who have completed at least one job in the last session. Within this group, the
-	/// ratio of payout to each validator is determined by the rewards assigned to the validator
-	/// compared to the total rewards in the system.
-	///
-	/// 2. The remaining fifty percent of the total rewards is divided among all validators. The
-	///    ratio for
-	/// each validator is determined by reading the storage `ValidatorRewardDistribution` that
-	/// assigns rewards based on the role type of active validators.
-	///
-	/// # Errors
-	///
-	/// Returns an error if the distribution encounters any issue while depositing rewards.
-	pub fn distribute_rewards() -> DispatchResult {
-		let current_era = CurrentEra::<T>::get().ok_or(Error::<T>::CannotGetCurrentEra)?;
-
-		let total_rewards = T::InflationRewardPerSession::get();
-
-		let active_validator_rewards: BTreeMap<_, _> =
-			Self::compute_active_validator_rewards(total_rewards / 2_u32.into());
-
-		let role_type_validator_rewards: BTreeMap<_, _> =
-			Self::compute_validator_rewards_by_role_type(total_rewards / 2_u32.into());
-
-		let mut combined_validator_rewards: BTreeMap<_, _> = Default::default();
-		for (validator, role_reward) in role_type_validator_rewards.iter() {
-			if let Some(reward) = active_validator_rewards.get(&validator) {
-				combined_validator_rewards
-					.insert(validator.clone(), role_reward.saturating_add(*reward));
-			} else {
-				combined_validator_rewards.insert(validator.clone(), *role_reward);
-			}
-		}
-
-		// record these rewards in the staking pallet storage
-		let mut era_reward_points = <ErasRewardPoints<T>>::get(&current_era);
-		let total_reward_points = total_rewards.saturating_add(era_reward_points.total.into());
-		era_reward_points.total = total_reward_points.try_into().unwrap_or(0_u32);
-
-		// add individual reward points to each eligible validator
-		for (validator, reward) in combined_validator_rewards.iter() {
-			let current_points = era_reward_points.individual.get(&validator).unwrap_or(&0_u32);
-			let new_points = reward.saturating_add((*current_points).into());
-			let new_points_as_u32: u32 = new_points.try_into().unwrap_or(0_u32);
-			era_reward_points.individual.insert(validator.clone(), new_points_as_u32.into());
-		}
-
-		ErasRewardPoints::<T>::insert(current_era, era_reward_points);
-		ValidatorJobsInEra::<T>::take();
-		Self::deposit_event(Event::<T>::RolesRewardPaid { total_rewards });
-
-		Ok(())
-	}
-
-	/// Computes the rewards for active validators based on the total rewards available
-	///
-	/// Given the total rewards available, this function calculates the rewards for each active
-	/// validator based on their share of the total rewards.
-	///
-	/// # Parameters
-	///
-	/// - `total_rewards`: Total rewards available for distribution among active validators.
-	///
-	/// # Returns
-	///
-	/// A `BTreeMap` containing the account IDs of active validators as keys and their respective
-	/// reward amounts as values.
-	pub fn compute_active_validator_rewards(
-		total_rewards: BalanceOf<T>,
-	) -> BTreeMap<T::AccountId, BalanceOf<T>> {
-		// Log total rewards available for debugging
-		log::debug!(
-			"Compute_active_validator_rewards, total rewards available {:?}",
-			total_rewards
-		);
-
-		// Retrieve active validators and their rewards from storage
-		let active_validators = ValidatorJobsInEra::<T>::get();
-
-		// Initialize a map to store rewards for active validators
-		let mut active_validator_reward_map: BTreeMap<_, _> = Default::default();
-
-		// Calculate the total jobs for all active validators
-		let total_jobs_completed: u32 =
-			active_validators.values().rfold(0_u32.into(), |acc, &x| acc + x);
-
-		// Distribute rewards to active validators based on their share
-		for (validator, jobs_completed) in active_validators.iter() {
-			// Calculate the share of rewards for the current validator
-			let validator_share = Perbill::from_rational(jobs_completed.clone(), total_jobs_completed);
-
-			// Calculate the actual reward amount for the current validator
-			let validator_share_of_total_reward = validator_share.mul_floor(total_rewards);
-			// Insert the validator's account ID and reward amount into the reward map
-			active_validator_reward_map
-				.insert(validator.clone(), validator_share_of_total_reward.saturating_add(*jobs_completed.into()));
-		}
-
-		// Return the map containing rewards for active validators
-		active_validator_reward_map
-	}
-
-	/// Computes the rewards for validators based on their role types and the total rewards
-	/// available
-	///
-	/// Given the total rewards available and the distribution of rewards among different role
-	/// types, this function calculates the rewards for validators based on their assigned role
-	/// types.
-	///
-	/// # Parameters
-	///
-	/// - `total_rewards`: Total rewards available for distribution among validators.
-	///
-	/// # Returns
-	///
-	/// A `BTreeMap` containing the account IDs of validators as keys and their respective
-	/// reward amounts as values.
-	///
-	/// # Remarks
-	///
-	/// The function retrieves the distribution of rewards among different role types from the
-	/// runtime configuration. It iterates over all accounts and their associated role types to
-	/// categorize validators into two groups:
-	/// - TSS validators
-	/// - ZkSaaS validators
-	/// The rewards are then calculated based on the distribution of rewards and the number of
-	/// validators in each group. The resulting rewards are stored in a map, where each validator's
-	/// account ID is associated with its reward amount.
-	pub fn compute_validator_rewards_by_role_type(
-		total_rewards: BalanceOf<T>,
-	) -> BTreeMap<T::AccountId, BalanceOf<T>> {
-		let mut tss_validators: Vec<T::AccountId> = Default::default();
-		let mut zksaas_validators: Vec<T::AccountId> = Default::default();
-		let dist = T::ValidatorRewardDistribution::get().get_reward_distribution();
-
-		// TODO : This is an unbounded query, potentially dangerous
-		for (acc, role_types) in AccountRolesMapping::<T>::iter() {
-			for role_type in role_types.iter() {
-				match role_type {
-					RoleType::Tss(_) =>
-						if !tss_validators.contains(&acc) {
-							tss_validators.push(acc.clone());
-						},
-					RoleType::ZkSaaS(_) =>
-						if !zksaas_validators.contains(&acc) {
-							zksaas_validators.push(acc.clone());
-						},
-					_ => (), // we dont care about other role types for inflation rewards
-				}
-			}
-		}
-
-		log::debug!("Found {:?} tss validators", tss_validators.len());
-		log::debug!("Found {:?} zksaas validators", zksaas_validators.len());
-
-		let tss_validator_reward = dist
-			.0
-			.mul_floor(total_rewards)
-			.checked_div(&(tss_validators.len() as u32).into())
-			.unwrap_or(Zero::zero());
-
-		let zksaas_validators_reward = dist
-			.1
-			.mul_floor(total_rewards)
-			.checked_div(&(zksaas_validators.len() as u32).into())
-			.unwrap_or(Zero::zero());
-
-		log::debug!("Reward per tss validator : {:?}", tss_validator_reward);
-		log::debug!("Reward per zksaas validator : {:?}", zksaas_validators_reward);
-
-		let mut validator_reward_map: BTreeMap<T::AccountId, BalanceOf<T>> = Default::default();
-
-		for validator in tss_validators {
-			validator_reward_map.insert(validator, tss_validator_reward);
-		}
-
-		for validator in zksaas_validators {
-			let default_balance: BalanceOf<T> = 0u32.into();
-			let current_rewards: BalanceOf<T> =
-				*validator_reward_map.get(&validator).unwrap_or(&default_balance);
-			validator_reward_map
-				.insert(validator, current_rewards.saturating_add(zksaas_validators_reward));
-		}
-
-		validator_reward_map
-	}
-
-	pub fn update_ledger_role_key(restaker: &T::AccountId, role_key: Vec<u8>) -> DispatchResult {
-		let mut ledger = Ledger::<T>::get(restaker).ok_or(Error::<T>::NoProfileFound)?;
-		let bounded_role_key: BoundedVec<u8, T::MaxKeyLen> =
-			role_key.try_into().map_err(|_| Error::<T>::KeySizeExceeded)?;
-		ledger.role_key = bounded_role_key;
-		Self::update_ledger(restaker, &ledger);
-		Ok(())
-	}
-}
-
 impl<T: Config> sp_runtime::BoundToRuntimeAppPublic for Pallet<T> {
 	type Public = T::RoleKeyId;
 }
@@ -577,6 +150,6 @@ impl<T: Config> OneSessionHandler<T::AccountId> for Pallet<T> {
 
 	// Distribute the inflation rewards
 	fn on_before_session_ending() {
-		let _ = Self::distribute_rewards();
+		// ignore
 	}
 }
